requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
Pillow>=10.0.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.9
tqdm>=4.65.0
PyYAML>=6.0
gradio>=4.0.0
matplotlib>=3.7.0
tensorboard>=2.14.0
scikit-image>=0.21.0
trimesh>=3.23.0
lpips>=0.1.4
pytorch-msssim>=1.0.0
requests>=2.31.0
gdown>=5.1.0
